refactor(vm): refactor state to its own file
